Fix overflowing number of arguments in XSCR;

+++ XtractQuery/Logic.Business.Level5ScriptManagement/Logic.Business.Level5ScriptManagement/Conversion/XscrCodeUnitConverter.cs
@@ -82,7 +82,7 @@ internal class XscrCodeUnitConverter : IXscrCodeUnitConverter
         result.Instructions.Add(new XscrScriptInstruction
         {
             Type = (short)instructionType,
-            ArgumentIndex = (short)argumentIndex,
+            ArgumentIndex = argumentIndex,
             ArgumentCount = (short)argumentCount
         });
     }
